+++ CHANGELOG.md
@@ -13,11 +13,17 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 1. Added support to Connection Templates Resource.
    - GET /rest/connection-templates/{id}
    - PUT /rest/connection-templates/{id}
+2. Added support to Appliance SNMPv1 Trap Destinations
+   - GET     /rest/appliance/trap-destinations/{id}
+   - POST    /rest/appliance/trap-destinations/{id}
+   - PUT     /rest/appliance/trap-destinations/{id}
+   - DELETE  /rest/appliance/trap-destinations/{id}
 
 ### Notes
 - This release supports API3000 minimally where we can use OneView v6.20 with this SDK.
 
 ### Oneview Features supported
+- Appliance SNMPv1 Trap Destinations
 - Connection Template
 
 # [v6.1.0-13]